+++ CHANGELOG.md
@@ -13,5 +13,5 @@
 * [Added] Data.init(hex:)
 * [Added] Base64.decode(data: Data) -> Data
 * [Added] Base64.decode(string: String) -> Data
-* [Added] Base64.base64.encode(data: Data) -> Data
+* [Added] Base64.standard.encode(data: Data) -> Data
 * [Added] Base64.urlSafe.encode(data: Data) -> Data